+++ src/components/diseases/Diseases.js
