+++ src/components/practices/PracticeDetails.js
@@ -9,8 +9,8 @@ class PracticeDetails extends Component {
         fontLoaded: false,
         title:'Practice Details'
       }
-   
-   
+
+
     render() {
         const { goBack } = this.props.navigation;
         return (
@@ -18,42 +18,42 @@ class PracticeDetails extends Component {
                 <Header style={[styles.headerStyle,styles.androidHeader]}>
                     <Left>
                         <Icon onPress={()=>this.props.navigation.navigate("Practices")} name="arrow-back" style={styles.leftIcon}/>
-                        
+
                     </Left>
                     <View style={{flex:1,marginLeft:5,justifyContent:'center'}}>
                         <Text style={{color:'white'}}>Practice Details</Text>
 
                     </View>
-                        
-                    
+
+
                     </Header>
                     <Content>
-                        
+
                         <View style={styles.ImageContainer}>
-                           <Image source={require('../assets/pfarm.jpg')} style={styles.imageStyl}  />
+                           <Image source={require('../../../assets/pfarm.jpg')} style={styles.imageStyl}  />
                            <View style={{position: 'absolute', top: 0, left: 5, right: 0, bottom: 0, justifyContent: 'center', alignItems: 'flex-start'}}>
                               <Text style={{color:'white',fontSize:18,fontWeight:'500'}}>Feeding-poultry</Text>
                             </View>
-                                      
-    
-                            
+
+
+
                          </View>
                          <View style={{paddingTop:5,paddingLeft:5,paddingRight:5}}>
                             <Text style={{flexShrink:1,flexWrap:'wrap'}}>
-                                Lorem ipsum dolor sit amet, consectetur adipiscing elit, 
-                                sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, 
-                                quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo 
+                                Lorem ipsum dolor sit amet, consectetur adipiscing elit,
+                                sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,
+                                quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo
                             </Text>
                          </View>
                          <View style={{flexDirection:'row',paddingTop:10,paddingLeft:5}}>
                              <View>
-                                <Image  style={{height: 100, width:100,}} source={require('../assets/poultry.jpg')} />
+                                <Image  style={{height: 100, width:100,}} source={require('../../../assets/poultry.jpg')} />
                              </View>
                              <Right style={{flex:1,paddingHorizontal:10,paddingRight:5}}>
                                 <Text style={{flexWrap:'wrap'}}>
-                                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, 
+                                    Lorem ipsum dolor sit amet, consectetur adipiscing elit,
                                     sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
-                                    quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo 
+                                    quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo
                                 </Text>
 
                              </Right>
@@ -61,9 +61,9 @@ class PracticeDetails extends Component {
                          </View>
                          <View style={{paddingTop:10,paddingLeft:5,paddingRight:5}}>
                             <Text style={{flexShrink:1,flexWrap:'wrap'}}>
-                                Lorem ipsum dolor sit amet, consectetur adipiscing elit, 
-                                sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, 
-                                quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo 
+                                Lorem ipsum dolor sit amet, consectetur adipiscing elit,
+                                sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,
+                                quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo
                             </Text>
                          </View>
                     </Content>
@@ -104,7 +104,7 @@ const styles={
         color:'white',
         marginRight: 10
     },
-    
+
     rightIcon:{
         color:"#fff"
 
@@ -119,7 +119,7 @@ const styles={
         flexDirection:'row',
         alignItems:'center',
         paddingHorizontal:5
-      
+
     },
     androidHeader:{
         ...Platform.select({
