+++ src/components/practices/Practices.js
@@ -15,7 +15,7 @@ class Practices extends Component {
           Roboto_medium: require("native-base/Fonts/Roboto_medium.ttf")
         });
         this.setState({ fontLoaded: true });
-        
+
       }
 
     render() {
@@ -24,7 +24,7 @@ class Practices extends Component {
                 <Header style={[styles.headerStyle,styles.androidHeader]}>
                     <Left>
                         <AntIcon onPress={()=>this.props.navigation.navigate("App")} name="left"  color="white" size={22}/>
-                    
+
                     </Left>
                     <Body>
                         {
@@ -32,90 +32,90 @@ class Practices extends Component {
                                 <Title style={{color:'white'}}>{this.state.title}</Title>
                             ) : null
                         }
-                        
+
                     </Body>
                     <Right>
                         <Icon name="search" style={styles.rightIcon}/>
                     </Right>
                 </Header>
                 <View style={{height:"10%",marginTop:5,flexDirection:'row', backgroundColor:'#f3f5f7',justifyContent:'space-between',marginLeft:5,marginRight:5,marginBottom:5}}>
-                    <View style={{flex:1,flexDirection:'row',justifyContent:'space-between',marginLeft:5,marginRight:5}}>                
+                    <View style={{flex:1,flexDirection:'row',justifyContent:'space-between',marginLeft:5,marginRight:5}}>
                             <Icon name="search"  style={{fontSize:20,paddingTop:17}}/>
-                            <Input style={{marginLeft:10}} placeholder="search..."/>                          
+                            <Input style={{marginLeft:10}} placeholder="search..."/>
                     </View>
-                    
+
                 </View>
                 <Content style={{backgroundColor:'#f3f5f7'}}>
                     <TouchableOpacity onPress={()=> this.props.navigation.navigate('PracticeDetails')}>
                          <Card style={{marginLeft:5,marginRight:5}}>
                             <View style={{paddingLeft:7,paddingTop:5}}>
-                                  
+
                                   <Text style={{fontWeight:'500',fontSize:16}}>Feeding</Text>
 
                               </View>
                             <CardItem style={{flexDirection:'column',flex:1}}>
-                                
+
                                 <View style={{flex:1,flexDirection:'row'}}>
-                                    <Image style={{borderRadius:9,width:90,height:90,resizeMode:'cover'}} source={require('../assets/feed.jpg')} />
+                                    <Image style={{borderRadius:9,width:90,height:90,resizeMode:'cover'}} source={require('../../../assets/feed.jpg')} />
                                     <Right style={{flex:1,alignItems:'flex-start',height:90,paddingHorizontal:9}}>
-                                        
+
                                             <Text style={{flexShrink:1,flexWrap:'wrap'}}>The best way to feed your Animals</Text>
-                                            
-                                        
+
+
                                     </Right>
-                                
+
                             </View>
-                                
+
                             </CardItem>
                         </Card>
 
                     </TouchableOpacity>
-                    
-                       
+
+
                         <Card style={{marginLeft:5,marginRight:5,marginTop:0}}>
                             <View style={{paddingLeft:7,paddingTop:5}}>
-                                  
+
                                   <Text style={{fontWeight:'500',fontSize:16}}>Breeding</Text>
 
                               </View>
                             <CardItem style={{flexDirection:'column',flex:1}}>
-                                
+
                                 <View style={{flex:1,flexDirection:'row'}}>
-                                    <Image style={{borderRadius:9,width:90,height:90,resizeMode:'cover'}} source={require('../assets/breeding.jpg')} />
+                                    <Image style={{borderRadius:9,width:90,height:90,resizeMode:'cover'}} source={require('../../../assets/breeding.jpg')} />
                                     <Right style={{flex:1,alignItems:'flex-start',height:90,paddingHorizontal:9}}>
-                                        
+
                                             <Text style={{flexShrink:1,flexWrap:'wrap'}}>The best way to breed your Animals</Text>
-                                            
-                                        
+
+
                                     </Right>
-                                
+
                             </View>
-                                
+
                             </CardItem>
                         </Card>
                         <Card style={{marginLeft:5,marginRight:5,marginTop:0}}>
                             <View style={{paddingLeft:7,paddingTop:5}}>
-                                  
+
                                   <Text style={{fontWeight:'500',fontSize:16}}>Housing</Text>
 
                               </View>
                             <CardItem style={{flexDirection:'column',flex:1}}>
-                                
+
                                 <View style={{flex:1,flexDirection:'row'}}>
-                                    <Image style={{borderRadius:9,width:90,height:90,resizeMode:'cover'}} source={require('../assets/housing.jpg')} />
+                                    <Image style={{borderRadius:9,width:90,height:90,resizeMode:'cover'}} source={require('../../../assets/housing.jpg')} />
                                     <Right style={{flex:1,alignItems:'flex-start',height:90,paddingHorizontal:9}}>
-                                        
+
                                             <Text style={{flexShrink:1,flexWrap:'wrap'}}>The best way to house your Animals</Text>
-                                            
-                                        
+
+
                                     </Right>
-                                
+
                             </View>
-                                
+
                             </CardItem>
                         </Card>
-                   
-                    
+
+
                 </Content>
             </Container>
         );
@@ -139,7 +139,7 @@ const styles={
         color:'white',
         marginRight: 10
     },
-    
+
     rightIcon:{
         color:"#fff"
 
@@ -154,7 +154,7 @@ const styles={
         flexDirection:'row',
         alignItems:'center',
         paddingHorizontal:5
-      
+
     },
     androidHeader:{
         ...Platform.select({
