+++ src/utils/AuthUtil.js
@@ -1,16 +1,14 @@
 import {AsyncStorage } from 'react-native'
 
-const checkUser=async(props)=>{
+const checkUser=async (props)=>{
     const bearerToken=await AsyncStorage.getItem('accessToken');
-    props.navigation.navigate(bearerToken?'App':'Auth')
-    console.log('Checking user')
-}
+    props.navigation.navigate(bearerToken?'App':'Auth');
+};
 
 const logout=async (props)=>{
-    await AsyncStorage.removeItem('accessToken')
+    await AsyncStorage.removeItem('accessToken');
     checkUser(props);
-    console.log('logging out')
-}
+};
 
 export {
     checkUser,
